requirements.txt
bs4>=0.0.2
python-dotenv>=0.9.9
fuzzywuzzy>=0.18.0
lxml>=6.0.2
openai>=2.6.1
pandas>=2.3.3
psycopg2>=2.9.11
pyarrow>=22.0.0
rapidfuzz>=3.14.3
requests>=2.32.5
tldextract>=5.3.0
tqdm>=4.67.1
warcio>=1.7.5
